Refactor les transactions pour pas raiser dedans

+++ zds/notification/models.py
@@ -4,15 +4,15 @@ from smtplib import SMTPException
 from django.contrib.auth.models import User
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
-from django.core.exceptions import ObjectDoesNotExist, MultipleObjectsReturned
 from django.core.mail import EmailMultiAlternatives
 from django.db import models, IntegrityError, transaction
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext_lazy as _
-
 from django.conf import settings
-from zds.forum.models import Topic
+from django.core.exceptions import ObjectDoesNotExist
+
 from zds.member.models import Profile
+from zds.forum.models import Topic
 from zds.notification.managers import NotificationManager, SubscriptionManager, TopicFollowedManager, \
     TopicAnswerSubscriptionManager, NewTopicSubscriptionManager
 from zds.utils.misc import convert_camel_to_underscore
@@ -106,7 +106,7 @@ class Subscription(models.Model):
         try:
             msg.send()
         except SMTPException:
-            pass
+            LOG.error('Failed sending mail for %s', self, exc_info=True)
 
     @staticmethod
     def has_read_permission(request):
@@ -131,19 +131,20 @@ class SingleNotificationMixin(object):
         assert hasattr(self, 'get_notification_url')
         assert hasattr(self, 'get_notification_title')
         assert hasattr(self, 'send_email')
-        with transaction.atomic():
-            if self.last_notification is None or self.last_notification.is_read:
+
+        if self.last_notification is None or self.last_notification.is_read:
+            with transaction.atomic():
+                notifications = list(Notification.objects.filter(subscription=self))
+                if len(notifications) > 1:
+                    LOG.error('Found %s notifications for %s', len(notifications), self, exc_info=True)
+                    Notification.objects.filter(pk__in=[n.pk for n in notifications[1:]]).delete()
+                    LOG.info('Duplicates deleted.')
+
                 # If there isn't a notification yet or the last one is read, we generate a new one.
                 try:
                     notification = Notification.objects.get(subscription=self)
                 except Notification.DoesNotExist:
                     notification = Notification(subscription=self, content_object=content, sender=sender)
-                except MultipleObjectsReturned:
-                    notifications = list(Notification.objects.filter(subscription=self))
-                    LOG.error('found %s notifications for %s', len(notifications), self, exc_info=True)
-                    Notification.objects.filter(pk__in=[n.pk for n in notifications[1:]]).delete()
-                    LOG.info('removed doubly.')
-                    notification = notifications[0]
                 notification.content_object = content
                 notification.sender = sender
                 notification.url = self.get_notification_url(content)
@@ -156,11 +157,11 @@ class SingleNotificationMixin(object):
 
                 if send_email and self.by_email:
                     self.send_email(notification)
-            elif self.last_notification is not None:
-                # Update last notification if the new content is older (case of unreading answer)
-                if not self.last_notification.is_read and self.last_notification.pubdate > content.pubdate:
-                    self.last_notification.content_object = content
-                    self.last_notification.save()
+        elif self.last_notification is not None and not self.last_notification.is_read:
+            # Update last notification if the new content is older (marking answer as unread)
+            if self.last_notification.pubdate > content.pubdate:
+                self.last_notification.content_object = content
+                self.last_notification.save()
 
     def mark_notification_read(self):
         """
@@ -210,27 +211,28 @@ class MultipleNotificationsMixin(object):
         if content is None:
             raise ValueError('Object content of notification must be defined')
 
-        with transaction.atomic():
-            content_notification_type = ContentType.objects.get_for_model(content)
-            notifications = list(Notification.objects.filter(subscription=self,
-                                                             content_type__pk=content_notification_type.pk,
-                                                             object_id=content.pk, is_read=False))
-            # handles cases where a same subscription lead to several notifications
-            if not notifications:
-                LOG.debug('nothing to mark as read')
-                return
-            elif len(notifications) > 1:
-                LOG.warning('%s notifications were find for %s/%s', len(notifications), content.type, content.title)
-                for notif in notifications[1:]:
-                    notif.delete()
+        try:
+            with transaction.atomic():
+                content_notification_type = ContentType.objects.get_for_model(content)
+                notifications = list(Notification.objects.filter(subscription=self,
+                                                                 content_type__pk=content_notification_type.pk,
+                                                                 object_id=content.pk, is_read=False))
+                # handles cases where a same subscription lead to several notifications
+                if not notifications:
+                    LOG.debug('nothing to mark as read')
+                    return
+                elif len(notifications) > 1:
+                    LOG.warning('%s notifications found for %s/%s', len(notifications), content.type, content.title)
+                    for notif in notifications[1:]:
+                        notif.delete()
 
-            notification = notifications[0]
-            notification.subscription = self
-            notification.is_read = True
-            try:
+                notification = notifications[0]
+                notification.subscription = self
+                notification.is_read = True
+
                 notification.save()
-            except IntegrityError:
-                LOG.exception('Could not save %s', notification)
+        except IntegrityError:
+            LOG.exception('Could not save %s', notification)
 
 
 class AnswerSubscription(Subscription):
